@@ -28,18 +28,19 @@
         <argLine>-Djava.security.egd=file:/dev/./urandom -Xmx1G</argLine>
         <m2e.apt.activation>jdt_apt</m2e.apt.activation>
         <run.addResources>false</run.addResources>
-        <jhipster-dependencies.version>8.0.0-beta.1</jhipster-dependencies.version>
-        <spring-boot.version>3.0.7</spring-boot.version>
+        <jhipster-dependencies.version>8.0.0-beta.2</jhipster-dependencies.version>
+        <spring-boot.version>3.1.1</spring-boot.version>
         <archunit-junit5.version>1.0.1</archunit-junit5.version>
-        <checkstyle.version>10.12.0</checkstyle.version>
+        <checkstyle.version>10.12.1</checkstyle.version>
         <checksum-maven-plugin.version>1.11</checksum-maven-plugin.version>
-        <frontend-maven-plugin.version>1.12.1</frontend-maven-plugin.version>
+        <frontend-maven-plugin.version>1.13.4</frontend-maven-plugin.version>
         <gatling-maven-plugin.version>4.3.7</gatling-maven-plugin.version>
         <gatling.version>3.9.5</gatling.version>
-        <git-commit-id-plugin.version>6.0.0</git-commit-id-plugin.version>
+        <git-commit-id-maven-plugin.version>6.0.0</git-commit-id-maven-plugin.version>
+        <h2.version>2.2.220</h2.version>
         <hibernate.version>6.1.7.Final</hibernate.version>
         <jacoco-maven-plugin.version>0.8.10</jacoco-maven-plugin.version>
-        <jaxb-runtime.version>4.0.2</jaxb-runtime.version>
+        <jaxb-runtime.version>4.0.3</jaxb-runtime.version>
         <jib-maven-plugin.architecture>amd64</jib-maven-plugin.architecture>
         <jib-maven-plugin.image>eclipse-temurin:17-jre-focal</jib-maven-plugin.image>
         <jib-maven-plugin.version>3.3.2</jib-maven-plugin.version>
